Split dir-mode cleanup helpers into their own file
skills_dir.go crossed the 500-line file limit the architecture test
enforces. The deletion-authorization and cleanup helpers form a
cohesive unit, so they moved verbatim to skills_dir_cleanup.go.
